Disable block based OTAs
The Google Play Edition has a smaller system partition, yet we use
size of the standard edition to generate block based OTAs. This
makes it impossible to install updates on the GPE, so disable block
based OTAs. In this way the size of the file system is determined at
runtime rather than build time.

Change-Id: I34f87e5500adfba9ae87a210c092ba1b17e04b56